pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-quality-intelligence"
version = "1.0.0"
description = "AI-powered code quality analysis tool with GitHub integration and interactive chatbot"
authors = [{name = "Code Quality Intelligence Team", email = "contact@codequalityintelligence.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["code quality", "static analysis", "AI", "LLM", "security", "complexity", "GitHub", "chatbot"]

dependencies = [
    "langchain==0.2.16",
    "langchain-groq==0.1.9", 
    "langchain-community==0.2.16",
    "click==8.1.7",
    "rich==13.7.1",
    "gitpython==3.1.43",
    "requests==2.31.0",
    "python-dotenv==1.0.1",
    "bandit==1.7.5",
    "radon==6.0.1",
    "safety==3.2.7",
    "pygments==2.17.2",
    "pathspec==0.12.1",
    "tqdm==4.66.2",
]

[project.optional-dependencies]
full = [
    "streamlit==1.31.0",
    "pandas==2.2.0", 
    "matplotlib==3.8.4",
    "plotly==5.19.0",
    "seaborn==0.13.2",
    "altair==5.2.0",
]
rag = [
    "chromadb==0.4.22",
    "sentence-transformers==2.2.2",
    "tiktoken==0.5.2",
    "faiss-cpu==1.8.0",
]
dev = [
    "pytest",
    "black", 
    "flake8",
    "mypy",
]

[project.scripts]
cqi = "code_quality_agent.cli:cli"
code-quality = "code_quality_agent.cli:cli"
code-quality-agent = "code_quality_agent.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/code-quality-intelligence"
"Bug Reports" = "https://github.com/yourusername/code-quality-intelligence/issues"
"Source Code" = "https://github.com/yourusername/code-quality-intelligence"
Documentation = "https://github.com/yourusername/code-quality-intelligence#readme"
